Version 2_UPDATED_ADC
Added in voltage regulators, power and programming headers, and bypass capacitors
TODO: Still need to add in SD card reader circuitry.

CHECK: Do we need to set AGND on the ADC to the sensor ground? Is the 0.5V regulator going to produce a steady reference voltage? What is the voltage for the strain gauges (UPDATE R2 VALUE WHEN WE KNOW)?